Added Ayu Dark theme
note: this is how the Ayu Dark theme appears in VS Code, minus one color. Run `colortest` inside the VS Code integrated terminal using the Ayu Dark theme and the top two text colors are a bit darker than the right column. I cannot fit that color into the palette so when you run colortest in Crosh both that top text color and the rightmost column are the same color.